Langchain/requirements.txt
#langchain framework
langchain
langchain-core

#openai Chatgpt
langchain-openai
openai

#Anthropics claude
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn